pyproject.toml
[tool.poetry]
name = "discord-bot-template"
version = "0.1.0"
description = ""
authors = ["sh1ma <in9lude@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {extras = ["voice"], version = "^1.4.1"}

[tool.poetry.dev-dependencies]
mypy = "^0.782"
black = "^20.8b1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
